refactor: AvatarImage의 props 정리, PrimaryAvatar로 개명
